fix: reduced the scope of the PR
