+++ docs/src/dev-guide/tooling-containers.md
@@ -3,15 +3,6 @@
 We publish (to [GitHub packages][gh-packages]) several Docker container images useful for building
 and running CLP:
 
-* Various manylinux_2_28 based images
-  ([x86_64][core-deps-manylinux_2_28_x86_64], [aarch64][core-deps-manylinux_2_28_aarch64])
-  containing the dependencies necessary to build CLP core in manylinux_2_28 environments
-
-  ```text
-  ghcr.io/y-scope/clp/clp-core-dependencies-manylinux_2_28_x86_64:main
-  ghcr.io/y-scope/clp/clp-core-dependencies-manylinux_2_28_aarch64:main
-  ```
-
 * An [image][core-deps-centos-stream-9] containing the dependencies necessary to build CLP core in a
   Centos Stream 9 x86 environment.
 
@@ -40,8 +31,6 @@ and running CLP:
   ghcr.io/y-scope/clp/clp-execution-x86-ubuntu-jammy:main
   ```
 
-[core-deps-manylinux_2_28_x86_64]: https://github.com/y-scope/clp/pkgs/container/clp%2Fclp-core-dependencies-manylinux_2_28_x86_64
-[core-deps-manylinux_2_28_aarch64]: https://github.com/y-scope/clp/pkgs/container/clp%2Fclp-core-dependencies-manylinux_2_28_aarch64
 [core-deps-centos-stream-9]: https://github.com/y-scope/clp/pkgs/container/clp%2Fclp-core-dependencies-x86-centos-stream-9
 [core-deps-ubuntu-jammy]: https://github.com/y-scope/clp/pkgs/container/clp%2Fclp-core-dependencies-x86-ubuntu-jammy
 [core-ubuntu-jammy]: https://github.com/y-scope/clp/pkgs/container/clp%2Fclp-core-x86-ubuntu-jammy
